Add Update operation. Fix Insert operation. Add tests.
The Insert operation was previously using the subscript behavior.
This meant object being inserted were not inserted but overwrote
the value at that index. In order to allow for proper behavior for
both subscripting and inserting and extra Operation of .Update was
added. Tests were augmented to ensure proper behavior.